apps/api/requirements.txt
# FastAPI et serveur
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Base de données
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
asyncpg==0.29.0

# CrewAI et IA
crewai==0.1.0
langchain==0.0.350
langchain-openai==0.0.2
langchain-community==0.0.1
openai==1.3.7

# Vector Database
qdrant-client==1.7.0

# PayloadCMS
payload==2.0.0

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Utilitaires
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
httpx==0.25.2
aiofiles==23.2.1

# Tests
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Linting et formatting
black==23.11.0
isort==5.12.0
flake8==6.1.0

# Monitoring
prometheus-client==0.19.0
structlog==23.2.0 